changed type for current_jb value to uint32

+++ include/baresip.h
@@ -1173,7 +1173,7 @@ void audio_level_put(const struct audio *au, bool tx, double lvl);
 int  audio_level_get(const struct audio *au, double *level);
 int  audio_debug(struct re_printf *pf, const struct audio *a);
 struct stream *audio_strm(const struct audio *au);
-uint16_t audio_jb_current_value(const struct audio* au);
+uint32_t audio_jb_current_value(const struct audio* au);
 int  audio_set_bitrate(struct audio *au, uint32_t bitrate);
 bool audio_rxaubuf_started(const struct audio *au);
 int  audio_start(struct audio *a);
